The future program
This is basically a program that sends through a network (VPN network also possible.) My personal goal is to turn on my pc from anywhere in the world without using port forwarding. this came with a few side effects, it became a window into my home network. 

This systems new goal will to become a full program that works to execute tasks within a network from a remote location. And this can be anything ofcourse